rebase galbi to kernel 6.9.1
tools: use basename to identify file in gen-mach-types

FILENAME is replaced by the full path to the executing script. If
the script is executed via a fully specified path, that is captured
in the output. Although it doesn't impact the output, it does trigger
reproducibility warnings/errors.

So we introduce a basename() function in the script and use it to
make sure the output file contains only the name of the awk script.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>